fix(velociraptor): move PersistenceSniper's mapper into its own module
The new mapper (added in a prior commit) pushed velociraptorImport.ts 4 lines
past the frozen size ledger from #384 (npm run check:size). Move
mapPersistenceSniper into its own analysis/persistenceSniperImport.ts, matching
the existing pattern for chainsawImport.ts/thorRowMap.ts, and share
mitreFromText via siemImport.ts instead of duplicating it. Behavior is
unchanged; register the new file in scripts/module-map.json (analysis/ingest).
